requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
boto3==1.38.46
    # via
    #   -r requirements.in
    #   django-storages
botocore==1.38.46
    # via
    #   boto3
    #   s3transfer
django==5.2.3
    # via
    #   -r requirements.in
    #   django-csp
    #   django-storages
django-csp==4.0
    # via -r requirements.in
django-storages[s3]==1.14.6
    # via -r requirements.in
dotenv==0.9.9
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
packaging==25.0
    # via django-csp
pillow==11.2.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.1
    # via dotenv
s3transfer==0.13.0
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
urllib3==2.5.0
    # via botocore
